Fix React import "fs" issue
