+++ src/index.node.ts
@@ -1,0 +1,4 @@
+import Aksharamukha from './aksharamukha/node';
+export default Aksharamukha;
+export * from './aksharamukha/node';
+export * from './enums';
